refactor: 타이머 로직 useCountdown 훅으로 분리
